+++ src/client/components/Login.tsx
@@ -5,6 +5,8 @@ import MuiAlert, { AlertProps } from '@material-ui/lab/Alert';
 
 import { usePlayerCount } from '../logic/stats';
 
+import packageJson from '../../../package.json';
+
 function Alert(props: AlertProps) {
   return <MuiAlert elevation={6} variant='filled' {...props} />;
 }
@@ -25,13 +27,15 @@ function Login({ handleLogin }: { handleLogin: (googleUser: any) => void }) {
 
   return (
     <div>
-      <div style={{ textAlign: 'center', maxWidth: '900px', margin: 'auto' }}>
+      <div style={{ textAlign: 'center', maxWidth: '900px', margin: 'auto', padding: '1em' }}>
         <div
           style={{
             textAlign: 'center'
           }}
         >
-          <h1 style={{ marginBottom: '0px' }}>Elemental Reborn</h1>
+          <h1 style={{ marginBottom: '0px' }}>
+            Elemental Reborn <span style={{ fontSize: '18px' }}>v{packageJson.version}</span>
+          </h1>
           <p style={{ marginTop: '0px' }}>{playerCount} Online</p>
         </div>
 
@@ -44,6 +48,18 @@ function Login({ handleLogin }: { handleLogin: (googleUser: any) => void }) {
           onSuccess={handleLogin}
           onFailure={onLoginError}
         />
+
+        <div style={{ textAlign: 'left' }}>
+          <h2>FAQ</h2>
+          <p>
+            <strong>Q</strong>: What happened to all the elements.
+          </p>
+          <p>
+            <strong>A</strong>: On 2020-09-18, I accidentally deleted all the recipes to the
+            elements with no way to get them back. The elements still existed, but none of them
+            would be creatable. The database was reset.
+          </p>
+        </div>
 
         <p style={{ fontSize: '12px' }}>
           This game is not officially affiliated with carykh's Elemental games, and none of your