+++ src/client/components/App.tsx
@@ -2,9 +2,9 @@ import React, { useEffect, useState } from 'react';
 
 import socket from '../socket';
 
-import { setMostRecentElements, setPlayerCount } from '../logic/stats';
+import { setElementCount, setMostRecentElements, setPlayerCount } from '../logic/stats';
 
-import { MostRecentElement } from '../../shared/types';
+import { ElementCount, MostRecentElement } from '../../shared/types';
 
 import Login from './Login';
 import Game from './Game';
@@ -22,13 +22,18 @@ function App() {
     function handleReceiveMostRecentElements(recent: MostRecentElement[]) {
       setMostRecentElements(recent);
     }
+    function handleReceiveElementCount(elementCount: ElementCount) {
+      setElementCount(elementCount);
+    }
 
     socket.on('player-count', handleReceivePlayerCount);
     socket.on('most-recent-elements', handleReceiveMostRecentElements);
+    socket.on('element-count', handleReceiveElementCount);
 
     return () => {
       socket.removeListener('player-count', handleReceivePlayerCount);
       socket.removeListener('most-recent-elements', handleReceiveMostRecentElements);
+      socket.removeListener('element-count', handleReceiveElementCount);
     };
   }, []);