+++ src/client/logic/stats.ts
@@ -6,7 +6,6 @@ import { MostRecentElement, ElementCount } from '../../shared/types';
 // Player count.
 export const playerCount = new SimpleState<number>(1);
 export const usePlayerCount = hookFromState(playerCount);
-
 export function setPlayerCount(newPlayerCount: number) {
   playerCount.set(newPlayerCount);
 }
@@ -14,9 +13,15 @@ export function setPlayerCount(newPlayerCount: number) {
 // Most recent elements.
 export const mostRecentElements = new SimpleState<MostRecentElement[]>(null as any);
 export const useMostRecentElements = hookFromState(mostRecentElements);
-
 export function setMostRecentElements(recent: any[]) {
   mostRecentElements.set(recent);
+}
+
+// Element count.
+export const elementCount = new SimpleState<ElementCount>(null as any);
+export const useElementCount = hookFromState(elementCount);
+export function setElementCount(newElementCount: ElementCount) {
+  elementCount.set(newElementCount);
 }
 
 export async function getElementCount(): Promise<ElementCount> {