+++ CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.3.0] - 2020-??-??
+## [1.4.0] - 2020-??-??
+### Added
+- Live element feed. See elements as soon as they are added.
+
+## [1.3.0] - 2020-09-20
 ### Added
 - Smooth scrolling when the game takes you to an element you created.
 - Elemental 4 implementation 👀.